+++ test/Altinn.Notifications.IntegrationTests/Notifications/TriggerController/Trigger_PastDueOrdersTests.cs
@@ -46,7 +46,7 @@ public class Trigger_PastDueOrdersTests : IClassFixture<IntegrationTestWebApplic
 
         // Assert
         string sql = $"select count(1) from notifications.orders where processedstatus = 'Processing' and alternateid='{orderId}'";
-        int actual = (int)await PostgreUtil.RunSqlReturnOutput<long>(sql);
+        long actual = await PostgreUtil.RunSqlReturnOutput<long>(sql);
 
         Assert.Equal(HttpStatusCode.OK, response.StatusCode);
         Assert.Equal(1, actual);
